var cleanup (#9)
* var cleanup

* remove branch reference in test

* remove ingress reference

* remove cert-mgr dependency fix sub deploy

* ref branch with fixes for mongo

* add storage class for test

* ingress add

* rm ccf reference

* fix inst and operator types

* remove dependency

* add depends

* fix cert output

* fix cert

* fix ca cert formatting

* fix ca cert format

* ca cert fix

* fix ca cert format

* ca output fix

* cert fix

* fix validate deploy and ca

* add sed spacing

* add deploy dependency

* fix ca formatting

* cluster ingress fix

* add license check

+++ .github/scripts/validate-deploy.sh
@@ -68,8 +68,8 @@ echo "Printing payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.
 cat "payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.yaml"
 
 
-
-sleep 7m
+# Deployment Validation
+sleep 5m
 count=0
 until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $count -eq 20 ]]; do
   echo "Waiting for namespace: ${NAMESPACE}"
@@ -86,23 +86,36 @@ else
 fi
 kubectl get all -n "${NAMESPACE}"
 
-#DEPLOYMENT="${INSTANCE_NAME}-${BRANCH}"
-#DEPLOYMENT="${COMPONENT_NAME}-${BRANCH}"
-#count=0
-#until kubectl get deployment "${DEPLOYMENT}" -n "${NAMESPACE}" || [[ $count -eq 20 ]]; do
-  #echo "Waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-  #count=$((count + 1))
-  #sleep 15
-#done
+count=0
+until kubectl get deployment ibm-truststore-mgr-controller-manager -n "${NAMESPACE}" || [[ $count -eq 10 ]]; do
+  echo "Waiting for deployment/ibm-truststore-mgr-controller-manager in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 60
+done
 
-#if [[ $count -eq 20 ]]; then
-  #echo "Timed out waiting for deployment/${DEPLOYMENT} in ${NAMESPACE}"
-  #kubectl get all -n "${NAMESPACE}"
-  #exit 1
-#fi
+if [[ $count -eq 10 ]]; then
+  echo "Timed out waiting for deployment/ibm-truststore-mgr-controller-manager in ${NAMESPACE}"
+  kubectl get all -n "${NAMESPACE}"
+  exit 1
+fi
+
+# Wait for license service to come online and obtain license ID
+sleep 5m
+
+count=0
+until kubectl get deployment ibm-sls-operator-instance-api-licensing -n "${NAMESPACE}" || [[ $count -eq 10 ]]; do
+  echo "Waiting for deployment/ibm-sls-operator-instance-api-licensing in ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 60
+done
+
+if [[ $count -eq 10 ]]; then
+  echo "Timed out waiting for deployment/ibm-sls-operator-instance-api-licensing in ${NAMESPACE}"
+  kubectl get all -n "${NAMESPACE}"
+  exit 1
+fi
 
 #kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
-
 
 cd ..
 rm -rf .testrepo
